+++ packages/perspective-workspace/src/ts/workspace/widget.ts
@@ -25,6 +25,7 @@ export class PerspectiveViewerWidget extends Widget {
     viewer: psp_viewer.HTMLPerspectiveViewerElement;
     _title: string;
     _is_table_loaded: boolean;
+    _is_pivoted: boolean;
     _restore_config?: () => Promise<void>;
     task?: Promise<void>;
 
@@ -33,6 +34,7 @@ export class PerspectiveViewerWidget extends Widget {
         this.viewer = viewer;
         this._title = "";
         this._is_table_loaded = false;
+        this._is_pivoted = false;
     }
 
     get name(): string {
@@ -65,16 +67,7 @@ export class PerspectiveViewerWidget extends Widget {
             this.viewer.setAttribute("table", table);
         }
 
-        // if (selectable) {
-        //     this.viewer.setAttribute("selectable", selectable);
-        // }
-
-        // if (editable) {
-        //     this.viewer.setAttribute("editable", editable);
-        // }
-
         const restore_config = () => this.viewer.restore({ ...viewerConfig });
-
         if (this._is_table_loaded) {
             return restore_config();
         } else {
@@ -88,14 +81,8 @@ export class PerspectiveViewerWidget extends Widget {
             table: this.viewer.getAttribute("table"),
         };
 
-        if (this.viewer.hasAttribute("selectable")) {
-            config.selectable = this.viewer.getAttribute("selectable");
-        }
-
-        if (this.viewer.hasAttribute("editable")) {
-            config.editable = this.viewer.getAttribute("editable");
-        }
-
+        delete config["theme"];
+        delete config["settings"];
         return config;
     }