+++ rust/perspective-viewer/test/js/save_restore.spec.js
@@ -142,4 +142,69 @@ test.describe("Save/Restore", async () => {
             "restore-restores-config-from-save.txt",
         ]);
     });
+
+    test("save/restore works in string format", async ({ page }) => {
+        const config = await page.evaluate(async () => {
+            const viewer = document.querySelector("perspective-viewer");
+            await viewer.getTable();
+            await viewer.restore({
+                settings: true,
+                group_by: ["State"],
+                columns: ["Profit", "Sales"],
+            });
+            return await viewer.save("string");
+        });
+
+        const config3 = await page.evaluate(async (config) => {
+            const viewer = document.querySelector("perspective-viewer");
+            await viewer.reset();
+            await viewer.restore(config);
+            return await viewer.save();
+        }, config);
+
+        expect(config3).toEqual({
+            ...DEFAULT_CONFIG,
+            columns: ["Profit", "Sales"],
+            plugin: "Debug",
+            group_by: ["State"],
+            settings: true,
+            theme: "Pro Light",
+        });
+
+        const contents = await get_contents(page);
+        await compareContentsToSnapshot(contents, [
+            "save-restore-works-in-string-format.txt",
+        ]);
+    });
+
+    test("save/restore works in arraybuffer format", async ({ page }) => {
+        const config3 = await page.evaluate(async () => {
+            const viewer = document.querySelector("perspective-viewer");
+            await viewer.getTable();
+            await viewer.restore({
+                settings: true,
+                group_by: ["State"],
+                columns: ["Profit", "Sales"],
+            });
+
+            const config = await viewer.save("arraybuffer");
+            await viewer.reset();
+            await viewer.restore(config);
+            return await viewer.save();
+        });
+
+        expect(config3).toEqual({
+            ...DEFAULT_CONFIG,
+            columns: ["Profit", "Sales"],
+            plugin: "Debug",
+            group_by: ["State"],
+            settings: true,
+            theme: "Pro Light",
+        });
+
+        const contents = await get_contents(page);
+        await compareContentsToSnapshot(contents, [
+            "save-restore-works-in-arraybuffer-format.txt",
+        ]);
+    });
 });